Setup for improved error handiling and extended chamnel_dependency
